Move EventJournalCacheEvent and EventJournalMapEvent back to public API
They are exposed as public API in jet.
