+++ hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/map/MapEventJournalReadTask.java
@@ -21,7 +21,7 @@ import com.hazelcast.client.impl.protocol.codec.MapEventJournalReadCodec;
 import com.hazelcast.instance.impl.Node;
 import com.hazelcast.map.impl.MapService;
 import com.hazelcast.map.impl.journal.MapEventJournalReadOperation;
-import com.hazelcast.map.impl.journal.EventJournalMapEvent;
+import com.hazelcast.map.journal.EventJournalMapEvent;
 import com.hazelcast.internal.nio.Connection;
 import com.hazelcast.nio.serialization.Data;
 import com.hazelcast.ringbuffer.impl.ReadResultSetImpl;
