+++ hazelcast/src/main/java/com/hazelcast/map/impl/journal/MapEventJournalFunctions.java
@@ -19,6 +19,7 @@ package com.hazelcast.map.impl.journal;
 import com.hazelcast.core.EntryEventType;
 import com.hazelcast.internal.journal.DeserializingEntry;
 import com.hazelcast.internal.serialization.SerializableByConvention;
+import com.hazelcast.map.journal.EventJournalMapEvent;
 
 import java.io.Serializable;
 import java.util.Map.Entry;
