+++ hazelcast/src/test/java/com/hazelcast/map/impl/journal/EventJournalMapEventAdapter.java
@@ -17,6 +17,7 @@
 package com.hazelcast.map.impl.journal;
 
 import com.hazelcast.journal.EventJournalEventAdapter;
+import com.hazelcast.map.journal.EventJournalMapEvent;
 
 public class EventJournalMapEventAdapter<K, V> implements EventJournalEventAdapter<K, V, EventJournalMapEvent<K, V>> {
     @Override
